Some logic needed for the pathing to work.
There's an issue with end turn behavior. Specifically, when a turn ends it doesn't allow the current user to move their units. It might have to do with how I'm checking the current unit's owner but if not I'm not sure. Secondly you can still end a turn before confirming a move which can do weird things but that should be an easy fix.