Fixed pathing missing a move
Oddly this should work fine without this fix. Essentially when a redundancyArray cell is <= validity it's either already been done or has been executed with a better value, thus validity >= redundancyArray should work. And yet it's mysteriously skipping a point when it isn't allowed to rerun equal runs. Either way, this should fix it.